requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.31.0
pandas>=2.1.4
numpy>=1.26.0
scikit-learn>=1.4.0
sentence-transformers>=2.3.1
beautifulsoup4>=4.12.3
lxml>=5.1.0
requests>=2.31.0
python-dotenv>=1.0.0
openpyxl>=3.1.2
torch>=2.1.2
google-generativeai>=0.8.0
